new adodb, ecommerce bug fixes, ebay categories bug fix

+++ libs/adodb/adodb-memcache.lib.inc.php
@@ -6,9 +6,12 @@ if (!defined('ADODB_DIR')) die();
 global $ADODB_INCLUDED_MEMCACHE;
 $ADODB_INCLUDED_MEMCACHE = 1;
 
+global $ADODB_INCLUDED_CSV;
+if (empty($ADODB_INCLUDED_CSV)) include(ADODB_DIR.'/adodb-csvlib.inc.php');
+
 /* 
 
-  V5.05 11 July 2008  (c) 2000-2008 John Lim (jlim#natsoft.com). All rights reserved.
+  V5.06 16 Oct 2008  (c) 2000-2009 John Lim (jlim#natsoft.com). All rights reserved.
   Released under both BSD license and Lesser GPL library license. 
   Whenever there is any discrepancy between the two licenses, 
   the BSD license will take precedence. See License.txt. 
@@ -70,24 +73,25 @@ $db->CacheExecute($sql);
 					$failcnt += 1;
 				}
 			}
-			if ($failcnt == sizeof($hosts)) {
+			if ($failcnt == sizeof($this->hosts)) {
 				$err = 'Can\'t connect to any memcache server';
 				return false;
 			}
-			
+			$this->_connected = true;
 			$this->_memcache = $memcache;
-			return 0;
+			return true;
 		}
 		
-		function writecache($filename, $contents,$debug, $secs2cache)
+		// returns true or false. true if successful save
+		function writecache($filename, $contents, $debug, $secs2cache)
 		{
 			if (!$this->_connected) {
 				$err = '';
 				if (!$this->connect($err) && $debug) ADOConnection::outp($err);
 			}
-			if ($this->_memcache) return false;
+			if (!$this->_memcache) return false;
 			
-			if (!$this->_memcache->set($filename, $contents, $this->compress, 0)) {
+			if (!$this->_memcache->set($filename, $contents, $this->compress, $secs2cache)) {
 				if ($debug) ADOConnection::outp(" Failed to save data at the memcached server!<br>\n");
 				return false;
 			}
@@ -95,18 +99,31 @@ $db->CacheExecute($sql);
 			return true;
 		}
 		
-		function &readcache($filename, &$err, $secs2cache, $rsClass)
+		// returns a recordset
+		function readcache($filename, &$err, $secs2cache, $rsClass)
 		{
+			$false = false;
 			if (!$this->_connected) $this->connect($err);
-			if ($this->_memcache) return false;
+			if (!$this->_memcache) return $false;
 			
 			$rs = $this->_memcache->get($filename);
 			if (!$rs) {
 				$err = 'Item with such key doesn\'t exists on the memcached server.';
 				return $false;
 			}
-	
-			$tdiff = intval($rs->timeCreated+$timeout - time());
+			
+			// hack, should actually use _csv2rs
+			$rs = explode("\n", $rs);
+            unset($rs[0]);
+            $rs = join("\n", $rs);
+ 			$rs = unserialize($rs);
+			if (! is_object($rs)) {
+				$err = 'Unable to unserialize $rs';		
+				return $false;
+			}
+			if ($rs->timeCreated == 0) return $rs; // apparently have been reports that timeCreated was set to 0 somewhere
+			
+			$tdiff = intval($rs->timeCreated+$secs2cache - time());
 			if ($tdiff <= 2) {
 				switch($tdiff) {
 					case 2: 
@@ -135,7 +152,7 @@ $db->CacheExecute($sql);
 				$err = '';
 				if (!$this->connect($err) && $debug) ADOConnection::outp($err);
 			}
-			if ($this->_memcache) return false;
+			if (!$this->_memcache) return false;
 			
 			$del = $this->_memcache->flush();
 			
@@ -148,6 +165,12 @@ $db->CacheExecute($sql);
 		
 		function flushcache($filename, $debug=false)
 		{
+			if (!$this->_connected) {
+  				$err = '';
+  				if (!$this->connect($err) && $debug) ADOConnection::outp($err); 
+			} 
+			if (!$this->_memcache) return false;
+  
 			$del = $this->_memcache->delete($filename);
 			
 			if ($debug)
